fix migration for compatible with redmine_2fa plugin

+++ db/migrate/002_add_mobile_phone_to_users.rb
@@ -1,9 +1,14 @@
 class AddMobilePhoneToUsers < ActiveRecord::Migration
   def up
-    add_column :users, :mobile_phone, :string
+    unless column_exists? :users, :mobile_phone
+      add_column :users, :mobile_phone, :string
+    end
+
   end
 
   def down
-    remove_column :users, :mobile_phone
+    unless Redmine::Plugin.installed?(:redmine_2fa)
+      remove_column :users, :mobile_phone
+    end
   end
 end
